[wip] restructure the project
