+++ includes/widgets-admin/elementor/registration_admin.php
@@ -5,12 +5,11 @@
  * @package ContentForms
  */
 
-namespace ThemeIsle\ContentForms\Includes\Widgets\Elementor\Registration;
+namespace ThemeIsle\ContentForms\Includes\Widgets_Admin\Elementor;
 
 use Elementor\Controls_Manager;
-use ThemeIsle\ContentForms\Includes\Widgets\Elementor\Elementor_Widget_Base;
 
-require_once TI_CONTENT_FORMS_PATH . '/includes/widgets/elementor/elementor_widget_base.php';
+require_once TI_CONTENT_FORMS_PATH . '/includes/widgets-admin/elementor/elementor_widget_base.php';
 
 /**
  * Class Registration_Admin
